refactor(shared): Refactor plugin helpers
- Move  into each plugin
