+++ packages/plugins/eslint-plugin-react-web-api/src/utils/create-rule.ts
@@ -1,4 +1,8 @@
-import { getDocsUrl } from "@eslint-react/shared";
+import { WEBSITE_URL } from "@eslint-react/shared";
 import { ESLintUtils } from "@typescript-eslint/utils";
 
-export const createRule = ESLintUtils.RuleCreator(getDocsUrl("web-api"));
+function getDocsUrl(ruleName: string) {
+  return `${WEBSITE_URL}/docs/rules/web-api-${ruleName}`;
+}
+
+export const createRule = ESLintUtils.RuleCreator(getDocsUrl);
