+++ packages/shared/src/config-adapters.ts
