+++ packages/shared/src/index.ts
@@ -1,9 +1,8 @@
 export * from "./_id";
 export * from "./_require";
+export * from "./config-adapters";
 export * from "./constants";
-export * from "./get-config-adapters";
-export * from "./get-doc-url";
-export * from "./get-react-version";
+export * from "./react-version";
 export * from "./regexp";
 export * from "./report";
 export * from "./settings";
